+++ CHANGES.md
@@ -16,6 +16,10 @@
     - Gestion des thèmes :
         - Bouton de réinitialisation avec confirmation
 
+## version 10.3.05
+- Correction :
+    - Thème : import d'un thème d'administration impossible.
+
 ## version 10.3.04
 - Corrections :
     - Warning lors d'une primo installation.